Remove convertThrowToError, rename ClosureWrapper
Because convertThrowToError cannot be made safe according to the letter
of the Swift runtime (it necessarily involves longjmp-ing across a
C->Swift call boundary). Instead, LuaClosure is given more focus in the
API, because lua_CFunction is less useful due to not being able to use
Lua errors.

At the same time, refactored all places that used or relied on
convertThrowToError by adding a set of helper functions to CLua's
extensions.c.