BREAKING CHANGE: LuaCallError no longer allows arbitrary Lua values
Because the `Error` protocol requires the struct to be `Sendable`, and
an error that potentially has a `LuaValue` definitely is not. And
Swift 6 rigidly enforces this meaning LuaSwift cannot compile as it
stands.

Therefore, LuaCallError now only works with strings, and if other values
need to be thrown and converted to Errors, a new setErrorConverter(_:)
API is added. Many API docs updated to reflect the fact that they may
throw more than just LuaCallError now.