Don't call PushableWithMetatable.metatable unnecessarily
checkRegistered was using it meaning the metatable would get recreated
multiple times every single time a value conforming to
PushableWithMetatable was pushed.

When metatables could be assumed to be `static let` constants it didn't
matter. With Swift 6 requiring them to be computed properties (thanks,
concurrency, I totally wanted all my declarations to be forced to be
'concurrency-safe') it does matter.

Since there's already an associatedtype that's the same as
metatable.type, the fix is straightforward.

This change also moves checkRegistered to use
`register(type:usingMetatable:)` instead of
`register(type:usingExistingMetatableFor:)` which removes another fetch
of the metatable (and correctly handles metatables in multi-level class
hierarchies).