pushMetatable(for:) should auto register if type conforms to PushableWithMetatable
Given that push(userdata:) does, pushMetatable(for:) should do the same
for consistency. The only way to do this appears to be to add another
overload of pushMetatable(for:). Fortunately the overload resolution
appears to do the right thing and call the right one depending on
whether the type conforms to PushableWithMetatable or not.